+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.2.0] 2025-08-01
+
+### Changed
+
+- ML_optimization: Better default values for operators of the Genetic Algorithm to encourage exploration.
+- ML_callbacks: `ModelCheckpoint` now accepts an optional checkpoint filename.
+- optimization_tools: `plot_optimal_feature_distributions()` automatically generates an output subdirectory to save plots.
+
 ## [6.1.2] 2025-08-01
 
 ### Fixed